+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Enhancements:
 * [Tests] Added StringAssert::isEmpty and StringAssert::isNotEmpty methods.
 * [Tests] Added CatchException::equalMessage method.
 * [Utilities] Added method Files::size.
+* [Functions] Added Functions::surroundWith.
 
 Bug fixes:
 * [Tests] Fixed StreamStub write method.